requirements.txt
streamlit
langchain
langchain-groq
python-docx